+++ tests/cli/test_config.py
@@ -1,3 +1,5 @@
+import unittest
+from pathlib import Path
 from typing import Any
 from unittest import TestCase
 
@@ -6,15 +8,13 @@ import fsspec
 import pytest
 
 from xrlint.cli.config import read_config
-from xrlint.cli.constants import CONFIG_DEFAULT_BASENAME
+from xrlint.cli.constants import DEFAULT_CONFIG_BASENAME
 from xrlint.config import Config
 from xrlint.config import ConfigList
 from xrlint.rule import RuleConfig
 
 yaml_text = """
 - name: yaml-test
-  plugins:
-    
   rules:
     rule-1: 2
     rule-2: "warn"
@@ -36,7 +36,7 @@ json_text = """
 """
 
 py_text = """
-def export_config(): 
+def export_configs(): 
     return [
         {
             "name": "py-test",
@@ -53,7 +53,7 @@ def export_config():
 # noinspection PyMethodMayBeStatic
 class CliConfigTest(TestCase):
     def test_read_config_yaml(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.yaml"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.yaml"
         with fsspec.open(config_path, mode="w") as f:
             f.write(yaml_text)
 
@@ -61,7 +61,7 @@ class CliConfigTest(TestCase):
         self.assert_config_ok(config, "yaml-test")
 
     def test_read_config_json(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.json"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.json"
         with fsspec.open(config_path, mode="w") as f:
             f.write(json_text)
 
@@ -69,7 +69,7 @@ class CliConfigTest(TestCase):
         self.assert_config_ok(config, "json-test")
 
     def test_read_config_python(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.py"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.py"
         with fsspec.open(config_path, mode="w") as f:
             f.write(py_text)
 
@@ -103,7 +103,7 @@ class CliConfigTest(TestCase):
             read_config(None)
 
     def test_read_config_with_format_error(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.json"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.json"
         with fsspec.open(config_path, mode="w") as f:
             f.write("{")
 
@@ -115,7 +115,7 @@ class CliConfigTest(TestCase):
             read_config(config_path)
 
     def test_read_config_with_unknown_format(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.toml"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.toml"
         with fsspec.open(config_path, mode="w") as f:
             f.write("")
 
@@ -126,25 +126,57 @@ class CliConfigTest(TestCase):
             read_config(config_path)
 
     def test_read_config_python_no_export(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.py"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.py"
         with fsspec.open(config_path, mode="w") as f:
             f.write("")
 
         with pytest.raises(
             click.ClickException,
-            match="memory://xrlint.config.py: missing definition of function 'export_config'",
+            match=(
+                "memory://xrlint.config.py: missing definition"
+                " of function 'export_configs'"
+            ),
         ):
             read_config(config_path)
 
     def test_read_config_with_exception(self):
-        config_path = f"memory://{CONFIG_DEFAULT_BASENAME}.py"
+        config_path = f"memory://{DEFAULT_CONFIG_BASENAME}.py"
         with fsspec.open(config_path, mode="w") as f:
-            f.write("def export_config():\n    raise ValueError('no config here!')\n")
+            f.write("def export_configs():\n    raise ValueError('no config here!')\n")
 
         from xrlint.util.importutil import UserCodeException
 
         with pytest.raises(
             UserCodeException,
-            match="while executing export_config\\(\\): no config here!",
+            match="while executing export_configs\\(\\): no config here!",
         ):
             read_config(config_path)
+
+
+class CliConfigResolveTest(unittest.TestCase):
+    def test_read_config_py(self):
+        self.assert_ok(
+            read_config(Path(__file__).parent / "configs" / "recommended.py")
+        )
+
+    def test_read_config_json(self):
+        self.assert_ok(
+            read_config(Path(__file__).parent / "configs" / "recommended.json")
+        )
+
+    def test_read_config_yaml(self):
+        self.assert_ok(
+            read_config(Path(__file__).parent / "configs" / "recommended.yaml")
+        )
+
+    def assert_ok(self, config_list: ConfigList):
+        self.assertIsInstance(config_list, ConfigList)
+        self.assertEqual(4, len(config_list.configs))
+        config = config_list.compute_config("test.zarr")
+        self.assertIsInstance(config, Config)
+        self.assertEqual("<computed>", config.name)
+        self.assertIsInstance(config.plugins, dict)
+        self.assertEqual({"xcube"}, set(config.plugins.keys()))
+        self.assertIsInstance(config.rules, dict)
+        self.assertIn("coords-for-dims", config.rules)
+        self.assertIn("xcube/cube-dims-order", config.rules)