+++ tests/plugins/xcube/test_plugin.py
@@ -18,8 +18,11 @@ class ExportPluginTest(TestCase):
         _plugin = export_plugin()
         self.assertEqual(
             {
+                "any-spatial-data-var",
                 "cube-dims-order",
+                "grid-mapping-naming",
                 "lat-lon-naming",
+                "single-grid-mapping",
             },
             set(_plugin.rules.keys()),
         )