+++ xrlint/cli/main.py
@@ -7,14 +7,14 @@ import click
 from xrlint.version import version
 from xrlint.cli.constants import DEFAULT_MAX_WARNINGS
 from xrlint.cli.constants import DEFAULT_OUTPUT_FORMAT
-from xrlint.cli.constants import CONFIG_DEFAULT_BASENAME
+from xrlint.cli.constants import DEFAULT_CONFIG_BASENAME
 
 
 @click.command(name="xrlint")
 @click.option(
     f"--no-default-config",
     "no_default_config",
-    help=f"Disable use of default configuration from {CONFIG_DEFAULT_BASENAME}.*",
+    help=f"Disable use of default configuration from {DEFAULT_CONFIG_BASENAME}.*",
     is_flag=True,
 )
 @click.option(
@@ -22,10 +22,30 @@ from xrlint.cli.constants import CONFIG_DEFAULT_BASENAME
     "-c",
     "config_path",
     help=(
-        f"Use this configuration, overriding {CONFIG_DEFAULT_BASENAME}.*"
+        f"Use this configuration, overriding {DEFAULT_CONFIG_BASENAME}.*"
         f" config options if present"
     ),
     metavar="PATH",
+)
+@click.option(
+    "--plugin",
+    "plugin_specs",
+    help=(
+        f"Specify plugins. MODULE is the name of Python module"
+        f" that defines an 'export_plugin()' function."
+    ),
+    metavar="MODULE",
+    multiple=True,
+)
+@click.option(
+    "--rule",
+    "rule_specs",
+    help=(
+        f"Specify rules. SPEC must have format"
+        f" '<rule-name>: <rule-config>' (note the space character)."
+    ),
+    metavar="SPEC",
+    multiple=True,
 )
 @click.option(
     "-f",
@@ -59,10 +79,12 @@ from xrlint.cli.constants import CONFIG_DEFAULT_BASENAME
 def main(
     no_default_config: bool,
     config_path: str | None,
+    plugin_specs: tuple[str, ...],
+    rule_specs: tuple[str, ...],
     max_warnings: int,
     output_format: str,
     output_file: str | None,
-    files: list[str] | None,
+    files: tuple[str, ...],
 ):
     """Lint the given dataset FILES.
 
@@ -76,22 +98,30 @@ def main(
     """
     from xrlint.cli.engine import CliEngine
 
+    if not files:
+        raise click.ClickException("No dataset files provided.")
+
     cli_engine = CliEngine(
         no_default_config=no_default_config,
         config_path=config_path,
+        plugin_specs=plugin_specs,
+        rule_specs=rule_specs,
         files=files,
         output_format=output_format,
         output_path=output_file,
     )
 
-    cli_engine.load_config()
-    results = cli_engine.verify_datasets()
+    config_list = cli_engine.load_config()
+    results = cli_engine.verify_datasets(config_list)
     report = cli_engine.format_results(results)
     cli_engine.write_report(report)
 
-    errors = sum(r.error_count for r in results)
-    warnings = sum(r.warning_count for r in results)
-    return 1 if errors > 0 or warnings > max_warnings else 0
+    error_status = sum(r.error_count for r in results) > 0
+    max_warn_status = sum(r.warning_count for r in results) > max_warnings
+    if max_warn_status and not error_status:
+        click.echo("maximum number of warnings exceeded.")
+    if max_warn_status or error_status:
+        raise click.exceptions.Exit(1)
 
 
 if __name__ == "__main__":  # pragma: no cover