+++ xrlint/config.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from xrlint.processor import ProcessorOp
 
 
-def get_base_config(recommended: bool = True):
+def get_core_config(recommended: bool = False):
     """Create a base configuration for the built-in plugins.
 
     Args:
@@ -28,26 +28,20 @@ def get_base_config(recommended: bool = True):
     Returns:
         A new `Config` object
     """
-    from xrlint.plugins.core import export_plugin as import_core_plugin
-    from xrlint.plugins.xcube import export_plugin as import_xcube_plugin
-
-    core_plugin = import_core_plugin()
-    xcube_plugin = import_xcube_plugin()
-
+    core_plugin = get_core_plugin()
     return Config(
         plugins={
             CORE_PLUGIN_NAME: core_plugin,
-            "xcube": xcube_plugin,
         },
-        rules=(
-            {
-                **core_plugin.configs["recommended"].rules,
-                **xcube_plugin.configs["recommended"].rules,
-            }
-            if recommended
-            else None
-        ),
+        rules=core_plugin.configs["recommended"].rules if recommended else None,
     )
+
+
+def get_core_plugin() -> "Plugin":
+    """Get the fully imported, populated core plugin."""
+    from xrlint.plugins.core import export_plugin
+
+    return export_plugin()
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -334,14 +328,8 @@ def merge_configs(
     config1: Config | dict[str, Any] | None,
     config2: Config | dict[str, Any] | None,
 ) -> Config:
-    if config1 is not None:
-        config1 = Config.from_value(config1)
-    else:
-        config1 = Config()
-    if config2 is not None:
-        config2 = Config.from_value(config2)
-    else:
-        config2 = Config()
+    config1 = Config.from_value(config1)
+    config2 = Config.from_value(config2)
     return config1.merge(config2)
 
 
@@ -363,20 +351,47 @@ class ConfigList:
         If `value` is already a `ConfigList` then it is returned as-is.
 
         Args:
-            value: A `ConfigList` object or `list` of values which can be
-                converted into `Config` objects.
+            value: A `ConfigList` object or `list` of items which can be
+                converted into `Config` objects including configuration
+                names of tyype `str`. The latter are resolved against
+                the plugin configurations seen so far in the list.
         Returns:
             A `ConfigList` object.
         """
         if isinstance(value, ConfigList):
             return value
-        if isinstance(value, list):
-            return ConfigList([Config.from_value(c) for c in value])
-        raise TypeError(
-            format_message_type_of(
-                "configuration list", value, "ConfigList|list[Config|dict]"
+        if not isinstance(value, list):
+            raise TypeError(
+                format_message_type_of(
+                    "configuration list", value, "ConfigList|list[Config|dict|str]"
+                )
             )
-        )
+
+        configs = []
+        plugins = {}
+        for item in value:
+            if isinstance(item, str):
+                plugin_name, config_name = (
+                    item.split("/", maxsplit=1)
+                    if "/" in item
+                    else (CORE_PLUGIN_NAME, item)
+                )
+                if CORE_PLUGIN_NAME not in plugins:
+                    plugins.update({CORE_PLUGIN_NAME: get_core_plugin()})
+                plugin: Plugin | None = plugins.get(plugin_name)
+                if (
+                    plugin is None
+                    or not plugin.configs
+                    or config_name not in plugin.configs
+                ):
+                    raise ValueError(f"configuration {item!r} not found")
+                config = Config.from_value(plugin.configs[config_name])
+            else:
+                config = Config.from_value(item)
+            configs.append(config)
+            plugins.update(config.plugins if config.plugins else {})
+
+        return ConfigList(configs=configs)
 
     def compute_config(self, file_path: str) -> Config | None:
         """Compute the configuration object for the given file path.