+++ xrlint/plugins/xcube/rules/lat_lon_naming.py
@@ -1,11 +1,10 @@
 from xrlint.node import DatasetNode
+from xrlint.plugins.xcube.constants import LAT_NAME, LON_NAME
 from xrlint.plugins.xcube.rules import plugin
 from xrlint.rule import RuleOp, RuleContext
 
-VALID_LON = "lon"
-VALID_LAT = "lat"
-INVALID_LONS = {"lng", "long", "longitude"}
-INVALID_LATS = {"ltd", "latitude"}
+INVALID_LAT_NAMES = {"ltd", "latitude"}
+INVALID_LON_NAMES = {"lng", "long", "longitude"}
 
 
 @plugin.define_rule(
@@ -13,22 +12,34 @@ INVALID_LATS = {"ltd", "latitude"}
     version="1.0.0",
     description=(
         f"Latitude and longitude coordinates and dimensions"
-        f" should be called {VALID_LAT!r} and {VALID_LON!r}."
+        f" should be called {LAT_NAME !r} and {LON_NAME !r}."
     ),
 )
 class LatLonNaming(RuleOp):
     def dataset(self, ctx: RuleContext, node: DatasetNode):
         lon_ok = _check(
-            ctx, "variable", node.dataset.variables.keys(), INVALID_LONS, VALID_LON
+            ctx, "variable", node.dataset.variables.keys(), INVALID_LON_NAMES, LON_NAME
         )
         lat_ok = _check(
-            ctx, "variable", node.dataset.variables.keys(), INVALID_LATS, VALID_LAT
+            ctx, "variable", node.dataset.variables.keys(), INVALID_LAT_NAMES, LAT_NAME
         )
         if lon_ok and lat_ok:
             # If variables have been reported,
             # we should not need to report (their) coordinates
-            _check(ctx, "dimension", node.dataset.sizes.keys(), INVALID_LONS, VALID_LON)
-            _check(ctx, "dimension", node.dataset.sizes.keys(), INVALID_LATS, VALID_LAT)
+            _check(
+                ctx,
+                "dimension",
+                node.dataset.sizes.keys(),
+                INVALID_LON_NAMES,
+                LON_NAME,
+            )
+            _check(
+                ctx,
+                "dimension",
+                node.dataset.sizes.keys(),
+                INVALID_LAT_NAMES,
+                LAT_NAME,
+            )
 
 
 def _check(ctx, names_name, names, invalid_names, valid_name):