+++ CHANGES.md
@@ -1,3 +1,15 @@
-## Version 0.0.1 (in development)
 
-Initial version. 
+## Early development snapshots
+
+- Version 0.0.2 (in development) 
+   - more rules
+   - more tests
+   - config list item can be config name
+   - xcube is no longer default plugin
+   - no rules by default
+   - no rules configured reports error
+   - CLI fails with max warnings exceeded 
+   - CLI fails with no files given
+   - new CLI options `--plugin` and `--rule`
+
+- Version 0.0.1 (30.12.2024) - Initial version.